requirements.txt
asgiref==3.4.1
Django==3.2.7
django-filter==2.4.0
django-phonenumber-field==5.2.0
djangorestframework==3.12.4
phonenumbers==8.12.31
pytz==2021.1
sqlparse==0.4.1
Pillow==8.3.2
django-mapbox-location-field==1.7.3

serializers~=0.2.4
defusedxml~=0.7.1
cffi~=1.14.6
future~=0.18.2
pip~=21.1.3
cryptography~=3.4.8
tornado~=6.1
mysqlclient~=2.0.3
PyYAML~=5.4.1
Jinja2~=3.0.1
pycparser~=2.20
setuptools~=56.0.0
tzdata~=2021.1
django-crispy-forms~=1.11.2
coreapi~=2.3.3
coreschema~=0.0.4
six~=1.13.0
uritemplate~=3.0.1
requests~=2.26.0